scrubbing_process/src/requirements.txt
pandas==2.1.4
openpyxl==3.1.2
pyyaml==6.0.1
openai==1.12.0
python-dotenv==1.0.0
tqdm==4.66.1
